% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nsclc}
\alias{nsclc}
\title{Non small cell lung cancer arm-based data}
\format{
A dataset.
}
\usage{
data(nsclc)
}
\description{
This dataset was used to demonstrate network meta-analysis with enrichment through weighting methods in the
paper Proctor et al.
It includes arm-based information on 14 studies with 3 different treatments and different covariates and binary outcome.
This dataset can be used to demonstrate the use of enriching-through-weighting methods
in BUGSnet.
}
\examples{
data(nsclc)

}
\references{
\insertRef{proctor22}{BUGSnet}
}
\keyword{datasets}
